Fall Cleaning
Going full shift into the job market trying to land a better job in a career more focused on my degree and what I want so removing the trash and rewriting resumes to fit what I am looking for.

Thanks to Network Chuck on Youtube for the inspiration and the assistance in setting this up.
https://www.youtube.com/watch?v=e2h_BreIxaQ&t=360s